+++ migration_wordpress/doc/changelog.rst
@@ -1,3 +1,7 @@
+v1.1.1
+======
+* Import users so blog posts are properly credited
+
 v1.1
 ====
 * Creates external id's to keep track of imported resources