+++ migration_wordpress_blog/doc/changelog.rst
@@ -1,3 +1,7 @@
+v1.0.2
+======
+* Import users so blog posts are properly credited
+
 v1.0.1
 ======
 * Bug fix related to initial import